Remove filters get on scrapping #15
