--- lib/game_board/zacatrus/domain/filters/zacatrus_filters.dart
@@ -1,9 +1,0 @@
-import 'package:zacatrusa/game_board/zacatrus/domain/filters/zacatrus_looking_for_filter_options.dart';
-
-class ZacatrusFilters {
-  const ZacatrusFilters({
-    this.lookingForOptions,
-  });
-
-  final ZacatrusLookingForFilterOptions? lookingForOptions;
-}
