+++ lib/game_board/zacatrus/domain/zacatrus_browse_page_data.dart
@@ -1,15 +1,11 @@
-import 'package:zacatrusa/game_board/zacatrus/domain/game_overview.dart';
-
-import 'filters/zacatrus_filters.dart';
+import 'game_overview.dart';
 
 class ZacatrusBrowsePageData {
   const ZacatrusBrowsePageData({
     required this.games,
-    this.filters,
     this.amount,
   });
 
   final int? amount;
   final List<GameOverview> games;
-  final ZacatrusFilters? filters;
 }
